requirements.txt
discord.py
google-generativeai>=0.8.3
fastapi
uvicorn
pydantic
python-dotenv
groq
supabase
playwright
python-dateutil
beautifulsoup4
lxml
httpx
pandas
tabulate
pytest
pytest-asyncio
slowapi
